requirements.txt
streamlit>=1.42.0
fastapi>=0.129.0
fpdf2>=2.7.0
google-genai>=1.62.0
httpx>=0.28.0
openai>=2.20.0
psycopg2-binary>=2.9.11
python-multipart>=0.0.22
requests>=2.32.0
resend>=2.23.0
stripe>=12.0.0
uvicorn>=0.40.0